Renames don't work if you only change case so some of the files are getting a stupid name temporarily.  Will be fixed next commit.
